fix more code smells
