+++ server/src/main/java/org/apache/iotdb/db/exception/query/LogicalOperatorException.java
@@ -31,6 +31,7 @@ public class LogicalOperatorException extends QueryProcessException {
     super("Error format in SQL statement, please check whether SQL statement is correct.",
         TSStatusCode.LOGICAL_OPERATOR_ERROR.getStatusCode());
   }
+
   public LogicalOperatorException(String message) {
     super(message, TSStatusCode.LOGICAL_OPERATOR_ERROR.getStatusCode());
   }
@@ -40,7 +41,7 @@ public class LogicalOperatorException extends QueryProcessException {
   }
 
   public LogicalOperatorException(String type, String message) {
-    super(String.format("Unsupported type: [%s]. " + message, type),
+    super(String.format("Unsupported type: [%s]. %s", type, message),
         TSStatusCode.LOGICAL_OPERATOR_ERROR.getStatusCode());
   }
 }
