+++ server/src/main/java/org/apache/iotdb/db/qp/Planner.java
@@ -19,7 +19,6 @@
 package org.apache.iotdb.db.qp;
 
 import java.time.ZoneId;
-import org.antlr.v4.runtime.misc.ParseCancellationException;
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.exception.query.LogicalOperatorException;
@@ -53,7 +52,7 @@ public class Planner {
   }
 
   public PhysicalPlan parseSQLToPhysicalPlan(String sqlStr)
-      throws QueryProcessException{
+      throws QueryProcessException {
     IoTDBConfig config = IoTDBDescriptor.getInstance().getConfig();
     return parseSQLToPhysicalPlan(sqlStr, config.getZoneID());
   }
