+++ server/src/main/java/org/apache/iotdb/db/query/externalsort/ExternalSortJobEngine.java
@@ -22,7 +22,7 @@ import java.io.IOException;
 import java.util.List;
 import org.apache.iotdb.tsfile.read.IPointReader;
 import org.apache.iotdb.db.query.reader.IReaderByTimestamp;
-import org.apache.iotdb.db.query.reader.chunkRelated.ChunkReaderWrap;
+import org.apache.iotdb.db.query.reader.chunkrelated.ChunkReaderWrap;
 
 
 public interface ExternalSortJobEngine {
