+++ server/src/main/java/org/apache/iotdb/db/query/reader/seriesrelated/SeriesRawDataPointReader.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.query.reader.seriesRelated;
+package org.apache.iotdb.db.query.reader.seriesrelated;
 
 import org.apache.iotdb.db.engine.querycontext.QueryDataSource;
 import org.apache.iotdb.db.query.context.QueryContext;
@@ -108,5 +108,6 @@ public class SeriesRawDataPointReader implements IPointReader {
 
   @Override
   public void close() throws IOException {
+    // no resources need to close
   }
 }
