+++ server/src/test/java/org/apache/iotdb/db/query/externalsort/ExternalSortEngineTest.java
@@ -29,7 +29,7 @@ import org.apache.iotdb.db.constant.TestConstant;
 import org.apache.iotdb.db.exception.StorageEngineException;
 import org.apache.iotdb.db.query.control.QueryResourceManager;
 import org.apache.iotdb.tsfile.read.IPointReader;
-import org.apache.iotdb.db.query.reader.chunkRelated.ChunkReaderWrap;
+import org.apache.iotdb.db.query.reader.chunkrelated.ChunkReaderWrap;
 import org.apache.iotdb.db.query.reader.universal.FakedSeriesReader;
 import org.apache.iotdb.db.query.reader.universal.PriorityMergeReader;
 import org.apache.iotdb.db.utils.EnvironmentUtils;
