+++ server/src/test/java/org/apache/iotdb/db/query/externalsort/FakeChunkReaderWrap.java
@@ -20,7 +20,7 @@ package org.apache.iotdb.db.query.externalsort;
 
 import org.apache.iotdb.tsfile.read.IPointReader;
 import org.apache.iotdb.db.query.reader.IReaderByTimestamp;
-import org.apache.iotdb.db.query.reader.chunkRelated.ChunkReaderWrap;
+import org.apache.iotdb.db.query.reader.chunkrelated.ChunkReaderWrap;
 
 /**
  * for test
